# docs/swagger.yaml
definitions:
  handler.ErrorResponse:
    properties:
      message:
        type: string
      type:
        type: string
    type: object
  handler.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/handler.ErrorResponse'
      success:
        type: boolean
    type: object
  handler.SuccessResponse:
    properties:
      data: {}
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /me:
    get:
      description: Требует access token в заголовке Authorization
      parameters:
      - default: Bearer <access_token>
        description: Bearer access_token
        in: header
        name: Authorization
        required: true
        type: string
      - default: Swagger-Test
        description: User-Agent
        in: header
        name: User-Agent
        type: string
      - default: 127.0.0.1
        description: IP адрес клиента
        in: header
        name: X-Forwarded-For
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Получить ID аутентифицированного пользователя
      tags:
      - auth
  /new_session/{user_id}:
    get:
      description: Возвращает access token в заголовке и refresh token в cookie
      parameters:
      - default: 0921ac27-5ec4-4031-a8f2-665e9c3c9eb3
        description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - default: Swagger-Test
        description: User-Agent
        in: header
        name: User-Agent
        type: string
      - default: 127.0.0.1
        description: IP адрес клиента
        in: header
        name: X-Forwarded-For
        type: string
      responses:
        "200":
          description: OK
          headers:
            Access-Token:
              description: Bearer <access_token>
              type: string
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Создать новую сессию
      tags:
      - auth
  /refresh:
    get:
      description: Требует access token в заголовке Authorization и cookie refresh_token.
      parameters:
      - default: Bearer <access_token>
        description: Bearer access_token
        in: header
        name: Authorization
        required: true
        type: string
      - default: Swagger-Test
        description: User-Agent
        in: header
        name: User-Agent
        type: string
      - default: 127.0.0.1
        description: IP адрес клиента
        in: header
        name: X-Forwarded-For
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Access-Token:
              description: Bearer <new_access_token>
              type: string
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Обновить access/refresh токены
      tags:
      - auth
  /refresh/revoke:
    post:
      description: Требует access token в заголовке Authorization.
      parameters:
      - default: Bearer <access_token>
        description: Bearer access_token
        in: header
        name: Authorization
        required: true
        type: string
      - default: Swagger-Test
        description: User-Agent
        in: header
        name: User-Agent
        type: string
      - default: 127.0.0.1
        description: IP адрес клиента
        in: header
        name: X-Forwarded-For
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Отозвать сессию
      tags:
      - auth
swagger: "2.0"